Variable Importance Plot of first output
The first script wants to specify which parameters are important for the biovolume estimation of the live phytoplankton cells. From the plot, the variable Channel 2 Red Fluoroscence Average has a high Mean Decrease in Accuracy, therefore it can be easily separated from the classification of the random forest. From the plot, the most important variable for biovolume estimation appears to be the Forward Scattering Average. 
All the variables are derived from flow cytometry using CytoSense tool.